Add setup project and a setupactions.dll to close the application (if running) during installation and deinstallation so that the executable can be removed.
Note that the current version gives a messagebox when shutdown. If the messagebox is not closed within a second, the executable cannot be removed after all. This should be fixed at a later time (i.e. implement custom popup which closes itself after a second or so, a-la customization dialog with 3 sec countdown).